pr-walkthrough: extract YAML→cards synthesis to pure shared module
Move validatePrWalkthroughYaml + mapYamlToWalkthroughPayload + DiffReferenceMapper
+ flattenDiffBlocks + helpers into src/shared/pr-walkthrough/yaml-to-cards.ts (pure,
no node:/server deps). walkthrough-yaml-schema.ts is now a thin re-export shim that
maps the structural result onto the server WalkthroughStorePayload. Behavior-preserving
(F-syn step 1).
